Initial duration of traffic lights now integer

+++ src/TrafficLight.java
@@ -3,9 +3,9 @@ import org.json.JSONObject;
 public class TrafficLight {
     private JSONObject JSONTrafficLight = new JSONObject();
     private char curr_color;
-    private double curr_duration; // TODO mudar para int
+    private int curr_duration;
 
-    TrafficLight(char color, double curr_duration) {
+    TrafficLight(char color, int curr_duration) {
         this.curr_color = color;
         this.curr_duration = curr_duration;
